API Parse get method returns error 400 "User not active" if user is not active.
wp_trksit_generate_url.php template on step 2 shows the error and the scrape fails.

TODO: Make more attractive things happen, like not show the form and show an error message with instructions how to reactivate.

TODO: Add this functionality to step 3 as well, and all the API calls that require active user.