settings page is only page available on plugin activation until user enters valid API keys.  Once API keys entered, the rest of the pages are shown.
wp_trksit_refreshToken() now sets a transient needed to check if the plugin is activated or not.  If the transient is not set, an API call goes out on page load to check that the token is valid.  The transient lives for a day so the plugin does not have to make an API call every time a page is loaded in the admin area.

TODO:  Add something at the API that when a user is disabled, the token can not refresh and invalidates all that user's tokens and sessions.